requirements.txt
streamlit>=1.24.1
langchain>=0.1.0
langchain-huggingface>=0.0.18
langchain-ollama>=0.0.8
langchain-community>=0.0.6
huggingface-hub>=0.16.4
python-docx>=0.8.12
pdfplumber>=0.9.0
faiss-cpu>=1.7.4
